msa-statservice/requirements.txt
Flask==0.10.1
kafka-python==1.3.3

